Downloads/副業/kintone/heroku-python/requirements.txt
achetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
urllib3==1.26.2